Revise npm install step
- No race conditions on restart
- Generate package-lock.json for service dependencies
- Auto-install node_modules for development
- Correctly use npm install and npm ci on first start and reload

Co-author @madnificent

+++ validate-package-json.sh
@@ -1,20 +1,24 @@
 #!/bin/bash
 cd /usr/src/app/
-node ./merge-package-json.js
-mv /tmp/merged-package.json /usr/src/app/app/package.json
+node ./validate-package-json.js
 cd /usr/src/app/app/
 
 ## Ensure package.json has module
-cat /usr/src/app/app/package.json | jq -e ".type" > /dev/null
-if [  $? -ne 0 ]
+if [ -f /usr/src/app/app/package.json ]
 then
+  cat /usr/src/app/app/package.json | jq -e ".type" > /dev/null
+  if [  $? -ne 0 ]
+  then
     echo '[WARNING] Adding "type": "module" to your package.json.'
     echo 'To remove this warning, add "type": "module" at the same level as "name" in your package.json'
     sed -i '0,/{/s/{/{\n  "type": "module",/' /usr/src/app/app/package.json
-else
+  else
     PACKAGE_TYPE=`cat /usr/src/app/app/package.json | jq -r ".type"`
     if [[ "$PACKAGE_TYPE" -ne "module" ]]
     then
-        echo '[WARNING] DIFFERENT TYPE THAN "module" IN package.json; CONTINUING WITH UNSPECIFIED BEHAVIOUR'
+      echo '[WARNING] DIFFERENT TYPE THAN "module" IN package.json; CONTINUING WITH UNSPECIFIED BEHAVIOUR'
     fi
+  fi
+else
+  echo '{ "type": "module" }' > /usr/src/app/app/package.json
 fi
